refactor(elements): Standardize SignIn context
