+++ packages/elements/src/react/sign-in/hooks/use-strategies.hook.ts
@@ -2,8 +2,8 @@ import type { SignInStrategy } from '@clerk/types';
 import { useCallback } from 'react';
 
 import { matchStrategy } from '~/internals/machines/utils/strategies';
-import type { SnapshotState } from '~/react/sign-in/contexts/machine.context';
-import { useSignInFlowSelector } from '~/react/sign-in/contexts/machine.context';
+import type { SnapshotState } from '~/react/sign-in/contexts/sign-in.context';
+import { SignInCtx } from '~/react/sign-in/contexts/sign-in.context';
 
 /**
  * Selects the Clerk current strategy
@@ -11,7 +11,7 @@ import { useSignInFlowSelector } from '~/react/sign-in/contexts/machine.context'
 const selector = (state: SnapshotState) => state.context.currentFactor?.strategy;
 
 export function useStrategies(_preferred?: SignInStrategy) {
-  const current = useSignInFlowSelector(selector);
+  const current = SignInCtx.useSelector(selector);
   const isActive = useCallback((name: string) => (current ? matchStrategy(current, name) : false), [current]);
 
   return {
