+++ packages/elements/src/react/sign-in/start.tsx
@@ -4,13 +4,13 @@ import type { PropsWithChildren } from 'react';
 
 import { Form } from '~/react/common/form';
 import { useActiveTags } from '~/react/hooks/use-active-tags.hook';
-import { useSignInFlow } from '~/react/sign-in/contexts/machine.context';
+import { SignInCtx } from '~/react/sign-in/contexts/sign-in.context';
 
 export type SignInStartProps = PropsWithChildren;
 
 export function SignInStart({ children }: SignInStartProps) {
-  const actorRef = useSignInFlow();
-  const activeState = useActiveTags(actorRef, 'state:start');
+  const ref = SignInCtx.useActorRef();
+  const activeState = useActiveTags(ref, 'state:start');
 
-  return activeState ? <Form flowActor={actorRef}>{children}</Form> : null;
+  return activeState ? <Form flowActor={ref}>{children}</Form> : null;
 }
